+++ src/Celeste.jl
@@ -6,12 +6,14 @@ require(joinpath(Pkg.dir("Celeste"), "src", "CelesteTypes.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "Transform.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "ModelInit.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "SDSS.jl"))
+require(joinpath(Pkg.dir("Celeste"), "src", "PSF.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "ElboDeriv.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "OptimizeElbo.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "Synthetic.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "SampleData.jl"))
 
 import SDSS
+import PSF
 import ElboDeriv
 import OptimizeElbo
 import ModelInit
